refactor(test_utils): consolidate environment variable test helpers and guards
- Removed `set_var_os` from environment module and tests.
- Replaced individual set/remove functions with `EnvGuard` struct that restores env vars on drop.
- Added `invalid_http_timeout_guard` to safely manage VK_HTTP_TIMEOUT test setup.
- Updated tests to use `EnvGuard` and `rstest` parametrized tests for clarity and safety.
- Centralized environment variable locking for thread-safe test setups.
